pyproject.toml
[project]
name = "claude-slop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "numpy>=2.3.3",
    "openai>=2.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
